+++ modules/hanna-react/src/ArticleCards.tsx
@@ -1,5 +1,4 @@
 import React from 'react';
-import type { Cleanup } from '@reykjavik/hanna-utils';
 
 import { CardList, ImageCardListProps, ImageCardProps } from './_abstract/_CardList.js';
 
@@ -8,9 +7,7 @@ export type ArticleCardsItemProps = ImageCardProps;
 /** @deprecated Use `ArticleCardsItemProps` instead (Remove in v0.11) */
 export type ArticleCardProps = ArticleCardsItemProps;
 
-export type ArticleCardsProps = Cleanup<
-  Pick<ImageCardListProps, 'cards' | 'imgPlaceholder'>
->;
+export type ArticleCardsProps = Pick<ImageCardListProps, 'cards' | 'imgPlaceholder'>;
 
 export const ArticleCards = (props: ArticleCardsProps) => (
   <CardList